Infdev 20100227-2

+++ src/main/java/net/minecraft/client/model/ModelRenderer.java
@@ -106,11 +106,11 @@ public final class ModelRenderer {
                     Vec3D var9 = var6.vertexPositions[1].vector3D.subtract(var6.vertexPositions[0].vector3D).normalize();
                     Vec3D var10 = var6.vertexPositions[1].vector3D.subtract(var6.vertexPositions[2].vector3D).normalize();
                     var9 = (new Vec3D(var9.yCoord * var10.zCoord - var9.zCoord * var10.yCoord, var9.zCoord * var10.xCoord - var9.xCoord * var10.zCoord, var9.xCoord * var10.yCoord - var9.yCoord * var10.xCoord)).normalize();
-                    var4.normal(-var9.xCoord, -var9.yCoord, -var9.zCoord);
+                    var4.normal((float)(-var9.xCoord), (float)(-var9.yCoord), (float)(-var9.zCoord));
 
                     for(int var11 = 0; var11 < 4; ++var11) {
                         PositionTextureVertex var12 = var6.vertexPositions[var11];
-                        var7.addVertexWithUV(var12.vector3D.xCoord * var8, var12.vector3D.yCoord * var8, var12.vector3D.zCoord * var8, var12.texturePositionX, var12.texturePositionY);
+                        var7.addVertexWithUV((float)var12.vector3D.xCoord * var8, (float)var12.vector3D.yCoord * var8, (float)var12.vector3D.zCoord * var8, var12.texturePositionX, var12.texturePositionY);
                     }
 
                     var4.draw();
