+++ src/main/java/net/minecraft/client/model/PositionTextureVertex.java
@@ -8,7 +8,7 @@ public final class PositionTextureVertex {
     public float texturePositionY;
 
     public PositionTextureVertex(float var1, float var2, float var3, float var4, float var5) {
-        this(new Vec3D(var1, var2, var3), var4, var5);
+        this(new Vec3D((double)var1, (double)var2, (double)var3), var4, var5);
     }
 
     public final PositionTextureVertex setTexturePosition(float var1, float var2) {
